linter warning

+++ src/solver/mip-based/VSSGenTimetableSolver_helper.cpp
@@ -201,8 +201,9 @@ bool cda_rail::solver::mip_based::VSSGenTimetableSolver::update_vss(
       (model->get(GRB_IntAttr_SolCount) >= 1)
           ? static_cast<size_t>(std::round(obj_ub - 1))
           : static_cast<size_t>(
-                std::ceil(iterative_factor * std::max(current_vss_number_e,
-                                                      static_cast<size_t>(1))));
+                std::ceil(iterative_factor *
+                          static_cast<double>(std::max(
+                              current_vss_number_e, static_cast<size_t>(1)))));
 
   if (target_vss_number_e >= vss_number_e) {
     target_vss_number_e = vss_number_e;
